syria_twitter/requirements.txt
aniso8601==1.1.0
Cython==0.23.4
Django==1.9.1
Flask==0.10.1
Flask-RESTful==0.3.4
gunicorn==19.3.0
image==1.4.4
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
numpy==1.10.4
Pillow==3.1.0
pymongo==3.1.1
python-bidi==0.4.0
python-dateutil==2.4.2
pytz==2015.7
requests==2.8.1
six==1.10.0
Werkzeug==0.11.2
wheel==0.24.0
wordcloud==1.1.3
